# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Kubico API
  version: 1.0.0
  description: API documentation for Kubico backend - Complete user and role management system
  contact:
    name: Kubico Support
    email: support@kubico.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.kubico.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated endpoints
  schemas:
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            email:
              type: string
              example: joao@example.com
            name:
              type: string
              example: João Silva
            role:
              type: string
              enum: [ADMIN, NORMAL, CLIENT, OWNER]
              example: NORMAL
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
        - userId
        - expiresAt
      properties:
        refreshToken:
          type: string
          description: Refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        userId:
          type: integer
          description: User ID
          example: 1
        expiresAt:
          type: string
          format: date-time
          description: Token expiration date
          example: 2026-03-16T12:00:00Z
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generates a new access token using a valid refresh token
      operationId: refresh
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized - invalid or expired refresh token
        '400':
          description: Invalid refresh token data
  /auth/auth/verify-email/{token}:
    get:
      tags:
        - Authentication
      summary: Verify email address
      description: Verifies the user's email address using the verification token sent via email
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Email verification token
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid or expired token
        '404':
          description: Token not found
...